+++ Common_glTF_Exporter/Common_glTF_Exporter.projitems
@@ -37,14 +37,16 @@
     <Compile Include="$(MSBuildThisFileDirectory)Export\GlbHeaderInfo.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Export\GltfJson.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Export\MeshOpt.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Materials\MaterialProperties.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Model\GlbComponents.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Model\Links.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Model\LoadContext.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Model\MaterialCacheDTO.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Model\Payload.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Model\RevitGridParametersObject.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Export\Compression.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Export\RevitGrids.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Export\RevitMaterials.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Materials\RevitMaterials.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Export\ZIP.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ExternalApplication.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ExternalCommand.cs" />
